Refactor to `HasContext` Concern
